add: model event and listener for async and sync

+++ dss-apps/dss-datamodel-center/dss-datamodel-indicator-management/src/main/java/com/webank/wedatasphere/dss/datamodel/indicator/service/IndicatorService.java
@@ -16,7 +16,7 @@ public interface IndicatorService extends IService<DssDatamodelIndicator> {
      * @param version
      * @return
      */
-    int addIndicator(IndicatorAddVO vo, String version) throws ErrorException;
+    Long addIndicator(IndicatorAddVO vo, String version) throws ErrorException;
